FASTAPI-MYSQL/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
cryptography
undetected-chromedriver
selenium
python-multipart
aiofiles
python-dotenv
pymysql
requests
faker